Move files to modules and components
