+++ packages/client-mobile/src/components/Map/index.tsx
@@ -3,7 +3,7 @@ import MapView from "react-native-maps";
 
 import styles from "./styles";
 import { Alert } from "react-native";
-import { useGeolocation } from "../geolocation";
+import { useGeolocation } from "../../modules/geolocation";
 
 // Melbourne, Victoria
 const defaultCoords = {
