+++ packages/client-mobile/src/components/NavBar/styles.ts
