+++ packages/client-mobile/src/components/WelcomeModal/styles.ts
