+++ packages/client-mobile/src/screens/Settings/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Text, View } from "react-native";
 
-import { ScreenProps } from "./types";
-import ScreenTemplate from "./ScreenTemplate";
+import { ScreenProps } from "../types";
+import ScreenTemplate from "../ScreenTemplate";
 
 const SettingsScreen = (_: ScreenProps<"settings">) => {
   return (
